+++ api/buildcraft/api/power/ILaserTarget.java
@@ -23,7 +23,7 @@ public interface ILaserTarget {
 	 *
 	 * @param energy
 	 */
-	void receiveLaserEnergy(double energy);
+	void receiveLaserEnergy(int energy);
 
 	/**
 	 * Return true if the Tile Entity object is no longer a valid target. For
